fastapi_worker/requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
PyPDF2
pinecone
psycopg2-binary
openai